Split whitespace only as much as we need to make tree builder progress
This saves almost 200 ms parsing 1MB of lorem ipsum with no tags.

We split whitespace at the beginning in order to ignore it while transitioning
from Initial to InBody mode.  But the bulk of the document is processed in
InBody mode and doesn't need split whitespace.  So it's inefficient to copy the
entire document into many separate tokens.

This also fixes a bug with multibyte characters.