Implement the "adoption agency algorithm"
This is a fairly straightforward and naive implementation that probably
has a lot of room for optimization. But it seems to match the spec's
intended behavior and passes a bunch of tests.

Implementing this exposed a bug in our handling of misnested <a> tags:
we were recording indices of <a> nodes in the list of active formatting
elements before running the adoption agency algorithm, then expecting
them still to be valid afterward. But the adoption agency algorithm can
mutate the list of active formatting elements, so we can't rely on the
indices anymore. Furthermore, the spec says that only a single <a> node
should be removed, which we now obey.