Don't allocate during rcdom::Node::drop
Instead of using Vec::extend with user-controlled piles of nodes, which
could trigger arbitrarily large heap allocations, use space that was
already allocated when constructing the tree: the parent pointers can
serve the same purpose as a recursion stack.

Also I think the previous implementation would remove all the children
of a node even if there were still other strong references to that node,
so this implementation should be more generally correct.