Remove rustc-test dependency (#528)
* xml5ever: remove unused rust-test dev-dependency

This isn't actually used by any code in the crate.

* rcdom: port tests/xml-tree-builder to a custom runner

This doesn't distribute the individual tests across the build-in
harness, hurting parallelism and process isolation. On the other
hand it's a minimal change to port off the rustc-test dependency
which hasn't been actively maintained and currently doesn't compile.

* rcdom: Port remaining tests to a custom runner

Move the custom test runner into its own `util::runner` module
and use it instead of rustc_test for the other test files.
This allows `cargo test` to complete under rust 1.77.1.

* github actions: Remove rustc-test/capture feature check

This is no longer available since the depdendency has been
removed.