Update to criterion 0.5 (#531)
Specify the current version of the benchmarking framework to address
cargo audit warnings about the dependencies of earlier releases.

Note that this requires rust 1.74.1 or later to run `cargo bench`
but MSRV (1.60.0) is only gated against the library target itself
and not dev-dependencies, so this is no change in the requirements
on downstream users.