ci: All jobs should pass for CI to pass
This change makes it so that CI does not pass unless all jobs defined in
the workflow file pass. This is necessary in order to prevent landing
changes that break CI jobs. `success()` and `failure()` which were used
before do not take into account dependent jobs.

This also fixes the "lint" job which was failing, by formatting the code.